Avoid local variable name collisions.
Scoping rules for local variables are a bit different in C#.
We have to keep track of local variables declared in parent
and previous child blocks. When a name collision is detected,
the variable is renamed. Also, allow specifying the expected
type of an expression being resolved. Depending on the
expected type, some conversion operations may need to be done.