Fix some method return type and interface issues
Java allows changing the return type of overriden methods.
That's not allowed in C#. When converting that type of code,
keep the return type of the base method and when the method
is called, do an explicit type cast.

Also, if a Java interface has fields or nested types, generate
it as a class.