Fix default mappings.
Removed some collections mappings which are now handled with extension
methods. Mapping Map.put, Map.get and other collection works for
simple conversions, but it doesn't work for Java code which subclasses
the Map class, or which depends on the specific behavior of those
methods (e.g. the put method must return the old value).
Convert.ToInt32 is used instead of parseInt because the former has an
overload that can take a radix, just like parseInt.